Update Header and Switch Multiplier Button
What did I do?

1. Header brand pushed to top left corner and navigation linked are moved
to the center.
2. The switch multiplier button now has a new design and only shows the
correct styling when that zone allows multiplier switching.
E.g. the switch multiplier button has no styling or functionality on the
"x1" or "x0" zones.

Why did I do it?

1. To make the header more visually appealing and user-friendly.
2. To improve the user experience by ensuring that the switch multiplier
button is only active in zones where it is applicable preventing
any confusion for the user.